+++ xbmc/cores/FFmpeg.cpp
@@ -114,7 +114,7 @@ void ff_avutil_log(void* ptr, int level, const char* format, va_list va)
 
   buffer += message;
   int pos, start = 0;
-  while( (pos = buffer.find_first_of('\n', start)) >= 0 )
+  while( (pos = (int)buffer.find_first_of('\n', start)) >= 0 )
   {
     if(pos>start)
       CLog::Log(type, "%s%s", prefix.c_str(), buffer.substr(start, pos-start).c_str());
